Extract hook to transition local video
